fix(i18n): address PR review feedback
- improve type safety in Settings.vue
- avoid v-html for translation hints
- centralize locale mapping logic in www app
- extract BCP 47 language tag mappings
- make internal language helpers private to the package
