+++ packages/languages/src/index.ts
@@ -1,102 +1,11 @@
+import { messages, type Locale } from './constants'
 import type { MessageKey } from './generated/keys'
+import { interpolate } from './interpolate'
+import { lookup } from './lookup'
 
-import en from './locales/en.json'
-import jp from './locales/jp.json'
-import zh from './locales/zh.json'
-
+export { defaultLocale, localeFlags, localeNames, locales, messages } from './constants'
+export type { Locale } from './constants'
 export type { MessageKey } from './generated/keys'
-
-// --- Shared Messages ---
-
-export const messages: Record<Locale, Record<string, string>> = { en, zh, jp }
-
-// --- Types ---
-
-export type Locale = 'en' | 'zh' | 'jp'
-export const locales = ['en', 'zh', 'jp'] as const satisfies readonly Locale[]
-export const defaultLocale: Locale = 'en'
-
-export const localeNames: Record<Locale, string> = {
-  en: 'English',
-  zh: '中文',
-  jp: '日本語',
-}
-
-export const localeFlags: Record<Locale, string> = {
-  en: '🇺🇸',
-  zh: '🇨🇳',
-  jp: '🇯🇵',
-}
-
-// --- Interpolation ---
-
-const TOKEN = /\{(\w+)\}/g
-
-export function interpolate(
-  template: string,
-  vars: Record<string, string | number | undefined | null>,
-): string {
-  return template.replace(TOKEN, (_, key: string) => {
-    const value = vars[key]
-    return value === undefined || value === null ? `{${key}}` : String(value)
-  })
-}
-
-// --- Lookup ---
-
-export function nodeDevFlag(): boolean {
-  const p = Reflect.get(globalThis, 'process') as { env?: Record<string, string | undefined> } | undefined
-  return p?.env?.DEV === 'true'
-}
-
-export function isDev(): boolean {
-  if (typeof import.meta !== 'undefined' && (import.meta as { env?: { DEV?: boolean } }).env?.DEV) {
-    return true
-  }
-  return nodeDevFlag()
-}
-
-/**
- * Lookup function that handles locale fallback and automatically checks
- * shared/common messages if the key is not found in the provided bundle.
- *
- * Priority:
- * 1. bundle[locale][key]
- * 2. shared messages[locale][key]
- * 3. bundle['en'][key] (fallback)
- * 4. shared messages['en'][key] (fallback)
- */
-export function lookup(
-  bundle: Record<Locale, Record<string, string>>,
-  locale: Locale,
-  key: string,
-): string {
-  // 1. Try local bundle for requested locale
-  const val = bundle[locale]?.[key]
-  if (val !== undefined)
-    return val
-
-  // 2. Try shared common messages for requested locale
-  const sharedVal = messages[locale]?.[key]
-  if (sharedVal !== undefined)
-    return sharedVal
-
-  // 3. Try en fallback (bundle then shared)
-  const fallback = bundle.en?.[key] || messages.en?.[key]
-  if (fallback !== undefined) {
-    if (isDev() && locale !== 'en') {
-      console.warn(
-        `[@typo/languages] Missing "${key}" in ${locale}, using en fallback`,
-      )
-    }
-    return fallback
-  }
-
-  if (isDev()) {
-    console.warn(`[@typo/languages] Missing key "${key}" in locale "${locale}"`)
-  }
-  return key
-}
 
 // --- Translators ---
